# spec/openapi.yaml
openapi: 3.1.0
info:
  title: PAL NAS Media Uploader API
  version: 1.0.0
  description: |
    Backend API for a self-hosted, multi-user NAS media library.

    ## Accounts and authorization

    The service has two roles. Administrators manage accounts but cannot use
    upload or gallery routes. Regular users upload into their configured folder,
    organize accessible media, and share owned media with other regular users.
    Resource lookup is scoped by the authenticated identity; inaccessible
    resources generally return `404` instead of revealing their existence.

    The first persistent administrator is created through the one-time HTML
    setup flow. Deployment bootstrap credentials authorize that operation only
    and are never stored as an application account. The initial administrator
    enrolls TOTP on first login. Every account subsequently created by an
    administrator also uses password plus TOTP.

    ## Sessions and browser access

    Authentication produces an opaque, HttpOnly cookie named
    `pal_medias_uploader_session`. Sessions last 30 days. Cross-origin browser
    requests must send credentials and their `Origin` must appear in
    `ENV_CORS_ALLOWED_ORIGINS`; an unlisted origin receives an empty `403`
    response. The cookie is `Secure` for HTTPS requests and uses `SameSite=None`
    only for secure cross-site requests; otherwise it uses `SameSite=Strict`.

    ## Uploads and media

...

...

    ## Albums

...
    have membership changed manually.

...

...

...

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /admin/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      description: Returns every configured account ordered by username. `totpEnabled` reflects whether an active TOTP secret is stored.
      responses:
        "200":
          description: All configured users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a TOTP-enabled user
      description: |
        Accepts a JSON body of at most 1 MiB and rejects unknown fields. The
        username is trimmed and case-sensitive. Roles are trimmed and normalized
        to lowercase. Passwords are bcrypt-hashed.

        A regular user requires a non-empty relative upload folder. Nested paths
        such as `family/alice` are allowed, while absolute paths and parent
        traversal are rejected. Administrator folders are always cleared.

        The service generates a TOTP secret for every managed account, including
        additional administrators. The provisioning URI and PNG data URI are
        returned only in this creation response and should be delivered securely.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created. The TOTP QR code is shown only in this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /media:
    get:
      tags: [Media]
      operationId: listMedia
      summary: List owned and shared media
      description: |
        In normal mode, returns active media owned by or shared with the current
        user. In trash mode, returns only trashed media owned by the current user;
        recipients never see an owner's trash. Favorite state is resolved for
        the requesting user, not globally.
      parameters:
        - name: search
          in: query
          description: Case-insensitive substring search over filename and MIME type.
          schema:
            type: string
        - name: kind
          in: query
          description: Filter by media kind. MIME types beginning with `video/` are videos; every other MIME type is treated as a photo.
          schema:
            type: string
            enum: [photo, video]
        - name: sort
          in: query
          description: Sort by completion time or filename-derived title.
          schema:
            type: string
            enum: [newest, oldest, title]
            default: newest
        - name: trash
          in: query
          description: When true, lists only the current user's trashed media.
          schema:
            type: boolean
            default: false
        - name: capturedAfter
          in: query
          description: Include media captured at or after this timestamp.
          schema: { type: string, format: date-time }
        - name: capturedBefore
          in: query
          description: Include media captured at or before this timestamp.
          schema: { type: string, format: date-time }
        - name: latitude
          in: query
          description: Center latitude for a radius search. Used only with longitude and radiusKm.
          schema: { type: number, format: double, minimum: -90, maximum: 90 }
        - name: longitude
          in: query
          description: Center longitude for a radius search. Used only with latitude and radiusKm.
          schema: { type: number, format: double, minimum: -180, maximum: 180 }
        - name: radiusKm
          in: query
          description: Maximum great-circle distance in kilometers.
          schema: { type: number, format: double, exclusiveMinimum: 0 }
      responses:
        "200":
          description: Accessible completed media.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Media"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /media/files/{id}/download:
    get:
      tags: [Media]
      operationId: downloadMedia
      summary: Download owned or shared media
      description: Returns an active media file only when the current user owns it or has a share. Filesystem paths are constrained to the configured media root.
      parameters:
        - $ref: "#/components/parameters/UploadID"
      responses:
        "200":
          description: Media bytes.
          headers:
            Content-Disposition:
              description: Attachment disposition containing the stored filename.
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/UploadNotFound"
  /media/files/{id}/shares:
    parameters:
      - $ref: "#/components/parameters/UploadID"
    post:
      tags: [Media]
      operationId: shareMedia
      summary: Share owned media with a regular user
      description: |
        Creates a library-wide share from the current user to the target user.
        The path media ID is used to verify that the current user owns active
        media; the grant is not limited to that item. The recipient immediately
        gains access to all of the owner's current active media, and future
        uploads become accessible without another share request.

        Shared media contributes to the recipient's lists, downloads, manual
        albums, favorites, storage totals, and next nightly automatic-album
        reconciliation. The target must be another regular user; administrators
        and the owner cannot be recipients. Repeating the relationship is
        idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareRequest"
      responses:
        "204":
          description: Library share added or already present.
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      tags: [Media]
      operationId: unshareMedia
      summary: Remove a library share
      description: |
        Removes the owner-to-recipient relationship represented by the owned
        media ID. All media belonging to that owner immediately becomes
        inaccessible to the recipient. It disappears from lists, downloads,
        storage totals, and album responses; future uploads are no longer shared.
        Ownership and the owner's albums are unchanged. Legacy clients may still
        send the username JSON body, but new clients should use the query
        parameter.
      parameters:
        - name: username
          in: query
          required: true
          description: Case-sensitive username of the regular-user recipient to remove.
          schema:
            type: string
            minLength: 1
      responses:
        "204":
          description: Library share removed.
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /admin/users/{id}/password:
    patch:
      tags: [Users]
      operationId: resetUserPassword
      summary: Reset a user's password and revoke their sessions
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "204":
          description: Password replaced and all target-user sessions deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /admin/users/{id}/trash-retention:
    patch:
      tags: [Users]
      operationId: setUserTrashRetention
      summary: Set a user's trash retention period
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrashRetentionRequest"
      responses:
        "204":
          description: Retention period updated.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /media/files/{id}/thumbnail:
    get:
      tags: [Media]
      operationId: downloadThumbnail
      summary: Get a generated thumbnail for accessible media
      parameters:
        - $ref: "#/components/parameters/UploadID"
      responses:
        "200":
          description: Generated JPEG thumbnail.
          headers:
            Cache-Control:
              schema: { type: string }
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/ResourceNotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pal_medias_uploader_session
      description: |
        Opaque session token stored only as a hash by the server. It is issued
        after successful TOTP verification, or immediately for the initial
        administrator, and expires after 30 days. Browser cookies are HttpOnly,
        use path `/`, and are Secure whenever the request is HTTPS.
  parameters:
    UserID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Target account UUID.
...
    UploadID:
      name: id
      in: path
      required: true
      description: 32-character lowercase hexadecimal identifier shared by an upload session and its completed media record.
      schema:
        type: string
        pattern: "^[0-9a-f]{32}$"
...
  schemas:
...
    User:
      type: object
      additionalProperties: false
      required: [id, username, role, totpEnabled, storageQuotaBytes, storageUsedBytes, trashRetentionDays]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          description: Unique, case-sensitive login name.
        role:
          type: string
          enum: [admin, user]
        uploadFolder:
          type: string
          description: Relative media destination. Empty for administrators.
        totpEnabled:
          type: boolean
          description: Whether an active TOTP secret is stored for this account.
        storageQuotaBytes:
          type: [integer, "null"]
          format: int64
          minimum: 1
          description: Per-user override or null when the environment default applies.
        storageUsedBytes:
          type: integer
          format: int64
          minimum: 0
        trashRetentionDays:
          type: integer
          minimum: 1
          maximum: 3650
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 1
          description: Unique, case-sensitive username; surrounding whitespace is removed.
        password:
          type: string
          format: password
          minLength: 12
        role:
          type: string
          enum: [admin, user]
          description: Trimmed and normalized to lowercase by the server.
        uploadFolder:
          type: string
          description: Required for regular users and ignored for administrators. Must be relative and cannot contain parent traversal.
    CreateUserResponse:
      type: object
      additionalProperties: false
      required: [user, provisioningUri, qrCode]
      properties:
        user:
          $ref: "#/components/schemas/User"
        provisioningUri:
          type: string
          format: uri
          description: Authenticator provisioning URI returned only when the account is created.
        qrCode:
          type: string
          pattern: "^data:image/png;base64,"
          description: PNG data URI for one-time TOTP setup, returned only when the account is created.
    ShareRequest:
      type: object
      additionalProperties: false
      required: [username]
      properties:
        username:
          type: string
          minLength: 1
          description: Case-sensitive username of a regular-user recipient.
...
    Media:
      type: object
      additionalProperties: false
      required: [id, ownerId, ownerUsername, title, kind, favorite, fileName, mimeType, fileSize, sha256, createdAt, shared]
      properties:
        id:
          type: string
          pattern: "^[0-9a-f]{32}$"
          description: Original completed upload identifier.
        ownerId:
          type: string
          format: uuid
        ownerUsername:
          type: string
        title:
          type: string
          description: Display title derived from the filename by removing its final extension.
        kind:
          type: string
          enum: [photo, video]
          description: "`video` when MIME type starts with `video/`; otherwise `photo`."
        favorite:
          type: boolean
          description: Personal state for the current user.
        fileName:
          type: string
          description: Sanitized basename supplied when the upload was created.
        mimeType:
          type: string
        fileSize:
          type: integer
          format: int64
          minimum: 1
        sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Lowercase digest computed by the server during completion.
        createdAt:
          type: string
          format: date-time
          description: Upload completion timestamp used for gallery ordering and automatic album dates.
        shared:
          type: boolean
          description: True when the current user is a recipient rather than the owner.
        deletedAt:
          type: string
          format: date-time
          description: Present only for owner trash listings.
        thumbnailUrl:
          type: string
          description: Protected thumbnail endpoint, present after processing succeeds.
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        durationSeconds:
          type: number
          format: double
          minimum: 0
        capturedAt:
          type: string
          format: date-time
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
...
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          format: password
          minLength: 12
...
    TrashRetentionRequest:
      type: object
      additionalProperties: false
      required: [days]
      properties:
        days:
          type: integer
          minimum: 1
          maximum: 3650
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          description: Stable or human-readable error message. Clients should primarily branch on HTTP status.
  responses:
    BadRequest:
      description: Malformed JSON or unknown authentication request field.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            malformed:
              value:
                error: Bad Request
...
    Forbidden:
      description: The authenticated account does not have the required role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UploadNotFound:
      description: Upload does not exist or is no longer active.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error: upload not found
    ResourceNotFound:
      description: The resource does not exist, is not owned by the current user, or is otherwise inaccessible. This avoids disclosing another user's resources.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...